Amends in response to code review
