+++ sso/localauth/urls.py
@@ -5,4 +5,5 @@ from sso.localauth import views
 
 urlpatterns = [
     url(r'^login/$', watch_login(views.LoginView.as_view()), name='login'),
+    url(r'^logout/$', views.session_logout, name='session-logout'),
 ]
